Move shared filter types to shared
