+++ api/src/types/services.ts
@@ -2,8 +2,7 @@ import { Knex } from 'knex';
 import { SchemaOverview } from '../types';
 import { Accountability } from '@directus/shared/types';
 import { Item, PrimaryKey } from './items';
-import { PermissionsAction } from './permissions';
-import { Query } from './query';
+import { Query, PermissionsAction } from '@directus/shared/types';
 
 export type AbstractServiceOptions = {
 	knex?: Knex;
