+++ api/src/utils/apply-query.ts
@@ -3,7 +3,8 @@ import { clone, get, isPlainObject, set } from 'lodash';
 import { customAlphabet } from 'nanoid';
 import validate from 'uuid-validate';
 import { InvalidQueryException } from '../exceptions';
-import { Aggregate, Filter, Query, Relation, SchemaOverview } from '../types';
+import { Relation, SchemaOverview } from '../types';
+import { Aggregate, Filter, Query } from '@directus/shared/types';
 import { applyFunctionToColumnName } from './apply-function-to-column-name';
 import { getColumn } from './get-column';
 import { getRelationType } from './get-relation-type';
