+++ api/src/utils/get-ast-from-query.ts
@@ -5,7 +5,8 @@
 import { Knex } from 'knex';
 import { cloneDeep, mapKeys, omitBy, uniq } from 'lodash';
 import { Accountability } from '@directus/shared/types';
-import { AST, FieldNode, NestedCollectionNode, PermissionsAction, Query, SchemaOverview } from '../types';
+import { AST, FieldNode, NestedCollectionNode, SchemaOverview } from '../types';
+import { Query, PermissionsAction } from '@directus/shared/types';
 import { getRelationType } from '../utils/get-relation-type';
 
 type GetASTOptions = {
