+++ api/src/utils/get-schema.ts
@@ -6,8 +6,8 @@ import { systemCollectionRows } from '../database/system-data/collections';
 import { systemFieldRows } from '../database/system-data/fields';
 import logger from '../logger';
 import { RelationsService } from '../services';
-import { Permission, SchemaOverview } from '../types';
-import { Accountability } from '@directus/shared/types';
+import { SchemaOverview } from '../types';
+import { Accountability, Permission } from '@directus/shared/types';
 import { toArray } from '@directus/shared/utils';
 import getDefaultValue from './get-default-value';
 import getLocalType from './get-local-type';
