+++ api/src/utils/merge-permissions.ts
@@ -1,5 +1,5 @@
 import { flatten, merge, omit } from 'lodash';
-import { Permission } from '../types';
+import { Permission, LogicalFilterOR } from '@directus/shared/types';
 
 export function mergePermissions(...permissions: Permission[][]): Permission[] {
 	const allPermissions = flatten(permissions);
@@ -29,7 +29,7 @@ function mergePerm(currentPerm: Permission, newPerm: Permission) {
 	if (newPerm.permissions) {
 		if (currentPerm.permissions && Object.keys(currentPerm.permissions)[0] === '_or') {
 			permissions = {
-				_or: [...currentPerm.permissions._or, newPerm.permissions],
+				_or: [...(currentPerm.permissions as LogicalFilterOR)._or, newPerm.permissions],
 			};
 		} else if (currentPerm.permissions) {
 			permissions = {
@@ -45,7 +45,7 @@ function mergePerm(currentPerm: Permission, newPerm: Permission) {
 	if (newPerm.validation) {
 		if (currentPerm.validation && Object.keys(currentPerm.validation)[0] === '_or') {
 			validation = {
-				_or: [...currentPerm.validation._or, newPerm.validation],
+				_or: [...(currentPerm.validation as LogicalFilterOR)._or, newPerm.validation],
 			};
 		} else if (currentPerm.validation) {
 			validation = {
