+++ api/src/utils/sanitize-query.ts
@@ -1,6 +1,7 @@
 import { flatten, get, merge, set } from 'lodash';
 import logger from '../logger';
-import { Aggregate, Filter, Meta, Query, Sort } from '../types';
+import { Meta } from '../types';
+import { Query, Sort, Aggregate, Filter } from '@directus/shared/types';
 import { Accountability } from '@directus/shared/types';
 import { parseFilter, deepMap } from '@directus/shared/utils';
 
@@ -44,7 +45,7 @@ export function sanitizeQuery(rawQuery: Record<string, any>, accountability?: Ac
 	}
 
 	if (rawQuery.meta) {
-		query.meta = sanitizeMeta(rawQuery.meta);
+		(query as any).meta = sanitizeMeta(rawQuery.meta);
 	}
 
 	if (rawQuery.search && typeof rawQuery.search === 'string') {
